pyproject.toml
[project]
name = "jordan-claw"
version = "0.1.0"
description = "Multi-tenant, multi-channel AI agent gateway"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-ai-slim[anthropic]>=1.75.0",
    "pydantic-settings>=2.5.0",
    "supabase>=2.11.0",
    "aiogram>=3.13.0",
    "tavily-python>=0.5.0",
    "structlog>=24.4.0",
    "httpx>=0.27.0",
    "caldav>=1.4.0",
    "croniter>=2.0.0",
    "python-frontmatter>=1.1.0",
    "openai>=1.50.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "logfire[fastapi,httpx]>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[project.scripts]
obsidian-sync = "scripts.obsidian_sync.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jordan_claw", "scripts"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src", "."]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
